+++ cpp/include/filehash.h
